Revert "chore: format using prettier"
This reverts commit 225ab5e498a464e45af35af202b40a27c649e3d2.

+++ contracts/farm/BeanBag.sol
@@ -8,11 +8,14 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 import "./interfaces/IBeanBag.sol";
 import "./interfaces/ILATTE.sol";
 
-contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
+contract BeanBag is ERC20('Bean Token', 'BEAN'), IBeanBag, Ownable {
+
   /// @notice latte token
   ILATTE public latte;
 
-  constructor(ILATTE _latte) public {
+  constructor(
+    ILATTE _latte
+  ) public {
     latte = _latte;
   }
 
@@ -29,17 +32,9 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
   /// @param _from The address of the account that will be debited
   /// @param _to The address of the account that will be credited
   /// @param _amount The amount to be moved
-  function transferFrom(
-    address _from,
-    address _to,
-    uint256 _amount
-  ) public virtual override returns (bool) {
+  function transferFrom(address _from, address _to, uint256 _amount) public virtual override returns (bool) {
     _transfer(_from, _to, _amount);
-    _approve(
-      _from,
-      _msgSender(),
-      allowance(_from, _msgSender()).sub(_amount, "BeanBag::transferFrom::transfer amount exceeds allowance")
-    );
+    _approve(_from, _msgSender(), allowance(_from, _msgSender()).sub(_amount, "BeanBag::transferFrom::transfer amount exceeds allowance"));
     _moveDelegates(_delegates[_from], _delegates[_to], _amount);
     return true;
   }
@@ -55,7 +50,7 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
   /// @notice Burn `_amount` BEAN to `_from`. Must only be called by MasterBarista.
   /// @param _from The address to burn BEAN from
   /// @param _amount The amount of BEAN that will be burned
-  function burn(address _from, uint256 _amount) external override onlyOwner {
+  function burn(address _from ,uint256 _amount) external override onlyOwner {
     _burn(_from, _amount);
     _moveDelegates(address(0), _delegates[_from], _amount);
   }
@@ -79,7 +74,7 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
   // https://github.com/compound-finance/compound-protocol/blob/master/contracts/Governance/Comp.sol
 
   /// @dev A record of each accounts delegate
-  mapping(address => address) internal _delegates;
+  mapping (address => address) internal _delegates;
 
   /// @notice A checkpoint for marking number of votes from a given block
   struct Checkpoint {
@@ -88,67 +83,88 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
   }
 
   /// @notice A record of votes checkpoints for each account, by index
-  mapping(address => mapping(uint32 => Checkpoint)) public checkpoints;
+  mapping (address => mapping (uint32 => Checkpoint)) public checkpoints;
 
   /// @notice The number of checkpoints for each account
-  mapping(address => uint32) public numCheckpoints;
+  mapping (address => uint32) public numCheckpoints;
 
   /// @notice The EIP-712 typehash for the contract's domain
-  bytes32 public constant DOMAIN_TYPEHASH =
-    keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
+  bytes32 public constant DOMAIN_TYPEHASH = keccak256("EIP712Domain(string name,uint256 chainId,address verifyingContract)");
 
   /// @notice The EIP-712 typehash for the delegation struct used by the contract
   bytes32 public constant DELEGATION_TYPEHASH = keccak256("Delegation(address delegatee,uint256 nonce,uint256 expiry)");
 
   /// @notice A record of states for signing / validating signatures
-  mapping(address => uint256) public nonces;
+  mapping (address => uint) public nonces;
 
   /// @notice An event thats emitted when an account changes its delegate
   event DelegateChanged(address indexed delegator, address indexed fromDelegate, address indexed toDelegate);
 
   /// @notice An event thats emitted when a delegate account's vote balance changes
-  event DelegateVotesChanged(address indexed delegate, uint256 previousBalance, uint256 newBalance);
+  event DelegateVotesChanged(address indexed delegate, uint previousBalance, uint newBalance);
 
   /**
-   * @notice Delegate votes from `msg.sender` to `delegatee`
-   * @param delegator The address to get delegatee for
-   */
-  function delegates(address delegator) external view returns (address) {
+  * @notice Delegate votes from `msg.sender` to `delegatee`
+  * @param delegator The address to get delegatee for
+  */
+  function delegates(address delegator)
+    external
+    view
+    returns (address)
+  {
     return _delegates[delegator];
   }
 
   /**
-   * @notice Delegate votes from `msg.sender` to `delegatee`
-   * @param delegatee The address to delegate votes to
-   */
+  * @notice Delegate votes from `msg.sender` to `delegatee`
+  * @param delegatee The address to delegate votes to
+  */
   function delegate(address delegatee) external {
     return _delegate(msg.sender, delegatee);
   }
 
   /**
-   * @notice Delegates votes from signatory to `delegatee`
-   * @param delegatee The address to delegate votes to
-   * @param nonce The contract state required to match the signature
-   * @param expiry The time at which to expire the signature
-   * @param v The recovery byte of the signature
-   * @param r Half of the ECDSA signature pair
-   * @param s Half of the ECDSA signature pair
-   */
+  * @notice Delegates votes from signatory to `delegatee`
+  * @param delegatee The address to delegate votes to
+  * @param nonce The contract state required to match the signature
+  * @param expiry The time at which to expire the signature
+  * @param v The recovery byte of the signature
+  * @param r Half of the ECDSA signature pair
+  * @param s Half of the ECDSA signature pair
+  */
   function delegateBySig(
     address delegatee,
-    uint256 nonce,
-    uint256 expiry,
+    uint nonce,
+    uint expiry,
     uint8 v,
     bytes32 r,
     bytes32 s
   ) external {
     bytes32 domainSeparator = keccak256(
-      abi.encode(DOMAIN_TYPEHASH, keccak256(bytes(name())), getChainId(), address(this))
+      abi.encode(
+        DOMAIN_TYPEHASH,
+        keccak256(bytes(name())),
+        getChainId(),
+        address(this)
+      )
     );
 
-    bytes32 structHash = keccak256(abi.encode(DELEGATION_TYPEHASH, delegatee, nonce, expiry));
+    bytes32 structHash = keccak256(
+      abi.encode(
+        DELEGATION_TYPEHASH,
+        delegatee,
+        nonce,
+        expiry
+      )
+    );
 
-    bytes32 digest = keccak256(abi.encodePacked("\x19\x01", domainSeparator, structHash));
+    bytes32 digest = keccak256(
+      abi.encodePacked(
+        "\x19\x01",
+        domainSeparator,
+        structHash
+      )
+    );
 
     address signatory = ecrecover(digest, v, r, s);
     require(signatory != address(0), "BeanBang::delegateBySig:: invalid signature");
@@ -158,23 +174,30 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
   }
 
   /**
-   * @notice Gets the current votes balance for `account`
-   * @param account The address to get votes balance
-   * @return The number of current votes for `account`
-   */
-  function getCurrentVotes(address account) external view returns (uint256) {
+  * @notice Gets the current votes balance for `account`
+  * @param account The address to get votes balance
+  * @return The number of current votes for `account`
+  */
+  function getCurrentVotes(address account)
+    external
+    view
+    returns (uint256)
+  {
     uint32 nCheckpoints = numCheckpoints[account];
     return nCheckpoints > 0 ? checkpoints[account][nCheckpoints - 1].votes : 0;
   }
 
   /**
-   * @notice Determine the prior number of votes for an account as of a block number
-   * @dev Block number must be a finalized block or else this function will revert to prevent misinformation.
-   * @param account The address of the account to check
-   * @param blockNumber The block number to get the vote balance at
-   * @return The number of votes the account had as of the given block
-   */
-  function getPriorVotes(address account, uint256 blockNumber) external view returns (uint256) {
+  * @notice Determine the prior number of votes for an account as of a block number
+  * @dev Block number must be a finalized block or else this function will revert to prevent misinformation.
+  * @param account The address of the account to check
+  * @param blockNumber The block number to get the vote balance at
+  * @return The number of votes the account had as of the given block
+  */
+  function getPriorVotes(address account, uint blockNumber)
+    external
+    view
+    returns (uint256) {
     require(blockNumber < block.number, "BeanBang::getPriorVotes:: not yet determined");
 
     uint32 nCheckpoints = numCheckpoints[account];
@@ -218,11 +241,7 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
     _moveDelegates(currentDelegate, delegatee, delegatorBalance);
   }
 
-  function _moveDelegates(
-    address srcRep,
-    address dstRep,
-    uint256 amount
-  ) internal {
+  function _moveDelegates(address srcRep, address dstRep, uint256 amount) internal {
     if (srcRep != dstRep && amount > 0) {
       if (srcRep != address(0)) {
         // decrease old representative
@@ -260,16 +279,14 @@ contract BeanBag is ERC20("Bean Token", "BEAN"), IBeanBag, Ownable {
     emit DelegateVotesChanged(delegatee, oldVotes, newVotes);
   }
 
-  function safe32(uint256 n, string memory errorMessage) internal pure returns (uint32) {
+  function safe32(uint n, string memory errorMessage) internal pure returns (uint32) {
     require(n < 2**32, errorMessage);
     return uint32(n);
   }
 
-  function getChainId() internal pure returns (uint256) {
+  function getChainId() internal pure returns (uint) {
     uint256 chainId;
-    assembly {
-      chainId := chainid()
-    }
+    assembly { chainId := chainid() }
     return chainId;
   }
 }
